{{cookiecutter.package_name}}/requirements.txt
pytorch
torchvision
tensorflow
numpy
cv2
matplotlib